requirements.txt
cffi==1.3.1
colorama==0.3.7
Django==1.9.5
djangorestframework==3.3.3
elasticsearch==2.3.0
elasticsearch-dsl==2.0.0
funcsigs==1.0.0
green==2.4.0
greenlet==0.4.9
mock==2.0.0
ordereddict==1.1
pbr==1.8.1
psycopg2==2.6.1
python-dateutil==2.5.2
python-termstyle==0.1.10
readline==6.2.4.1
six==1.10.0
urllib3==1.15
